Restructure web-version into backend and frontend
