+++ web-version/.gitignore
@@ -1,2 +1,2 @@
 .DS_Store
-downloads/
+/downloads/
