+++ web-version/backend/src/history-page-assets.ts
