+++ web-version/backend/src/http/handlers/assets.ts
