+++ web-version/backend/src/http/routes/assets/index.ts
