+++ web-version/frontend/src/history-page.css
