+++ web-version/frontend/src/history-page/client/components/PlaybackWindow.tsx
