+++ web-version/frontend/src/history-page/client/hooks/useBusyMap.ts
