+++ web-version/frontend/src/history-page/client/hooks/useDialogState.ts
