+++ web-version/frontend/src/history-page/client/hooks/useHistoryWebSocket.ts
