+++ web-version/frontend/src/history-page/client/types.ts
