+++ web-version/frontend/src/history-page/client/utils/dom.ts
