+++ web-version/frontend/src/history-page/client/utils/format.ts
