+++ web-version/frontend/src/history-page/components/ProgressCell.tsx
