+++ web-version/frontend/src/history-page/components/SearchForm.tsx
