+++ web-version/frontend/src/history-page/components/TableRow.tsx
