+++ web-version/frontend/src/history-page/shared/types.ts
@@ -1,0 +1,5 @@
+export type {
+    FormatOption,
+    AddDownloadDialogMode,
+    HistoryItemBase
+} from '@shared/history-page';
