+++ web-version/frontend/src/history-page/types.ts
@@ -1,0 +1,6 @@
+export type {
+    HistoryItem,
+    PaginationState,
+    HistoryPageViewProps,
+    FormatOption
+} from '@shared/history-page';
