+++ web-version/frontend/src/history-page/utils/format.ts
