+++ web-version/frontend/src/history-page/utils/links.ts
