+++ web-version/frontend/src/history-page/utils/status.ts
