--- web-version/src/backend/index.ts
@@ -1,1 +1,0 @@
-export {};
