+++ web-version/tsconfig.json
@@ -4,17 +4,22 @@
     "module": "nodenext",
     "moduleResolution": "nodenext",
     "jsx": "react-jsx",
-    "outDir": "dist",
-    "rootDir": "src",
+    "outDir": "dist/backend",
+    "rootDir": "backend/src",
+    "baseUrl": ".",
+    "paths": {
+      "@shared/history-page": ["shared/src/history-page.d.ts"],
+      "@shared/*": ["shared/src/*"]
+    },
     "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true,
     "resolveJsonModule": true,
     "declaration": true,
-    "types": ["node", "react"],
-    "typeRoots": ["./src/types", "./node_modules/@types"]
+    "types": ["node"],
+    "typeRoots": ["./backend/src/types", "./node_modules/@types"]
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.d.ts"],
-  "exclude": ["dist"]
+  "include": ["backend/src/**/*.ts", "backend/src/**/*.tsx", "backend/src/**/*.d.ts"],
+  "exclude": ["dist", "frontend"]
 }
